requirements.txt
Django==1.6.5
python-ldap==2.4.15
